backend/requirements.txt
django
djangorestframework
django-cors-headers
python-dotenv
requests
psycopg2-binary
whitenoise
dj-database-url
